Rename 'translation-cleanup' directory to 'translations-cleanup' for consistency across imports and file paths.
