+++ src/translations-cleanup/fileScanner.ts
