+++ src/translations-cleanup/patterns.ts
